# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Valuz Agent API
  version: "0.3.0"
  description: Valuz Agent API contract — single source of truth
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /v1/agents:
    get:
      summary: List official Agents (read-only)
      operationId: listAgents
      tags: [agents]
      parameters:
        - name: source
          in: query
          required: false
          schema:
            type: string
          description: Filter by source (official|custom).
      responses:
        "200":
          description: Official agents
          content:
            application/json:
              schema:
                type: object
                required: [agents]
                properties:
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Agent"
    post:
      summary: Create a user-defined Agent
      operationId: createAgent
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "409":
          description: Slug already exists
  /v1/agents/{slug}:
    get:
      summary: Get a single Agent
      operationId: getAgent
      tags: [agents]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Agent not found
    patch:
      summary: Update an Agent
      operationId: updateAgent
      tags: [agents]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentRequest"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Agent not found
    delete:
      summary: Delete an Agent
      operationId: deleteAgent
      tags: [agents]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: cascade
          in: query
          required: false
          description: >-
            When true, first 解除 every 派驻 (project membership) the agent has,
            then delete it — the confirmed-delete path. When false (default), an
            agent still deployed to a project returns 409.
          schema:
            type: boolean
            default: false
      responses:
        "204":
          description: Deleted
        "404":
          description: Agent not found
        "409":
          description: Agent still deployed to one or more projects (解除派驻 first; pass cascade=true to force)
  /v1/agents/{slug}/deployments:
    get:
      summary: List the projects an agent is deployed (派驻) into
      operationId: listAgentDeployments
      tags: [agents]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deployments (live-reference 派驻 across projects)
          content:
            application/json:
              schema:
                type: object
                required: [deployments, count]
                properties:
                  count: {type: integer}
                  deployments:
                    type: array
                    items:
                      type: object
                      required: [project_id, agent_slug]
                      properties:
                        project_id: {type: string}
                        agent_slug:
                          type: string
                          description: Project-local member handle.
        "404":
          description: Agent not found
  /v1/projects/{project_id}/agents:
    get:
      summary: List project member agents
      operationId: listProjectAgents
      tags: [agents]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project member agents
          content:
            application/json:
              schema:
                type: object
                required: [agents]
                properties:
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/MemberWithAgent"
    post:
      summary: Create a blank (template-free) agent in a project
      operationId: createBlankAgent
      tags: [agents]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlankAgentRequest"
      responses:
        "201":
          description: Created member agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberWithAgent"
        "409":
          description: Agent slug already exists in project
  /v1/projects/{project_id}/agents:deploy:
    post:
      summary: Deploy (派驻, live-reference) a library agent into a project
      operationId: deployAgent
      tags: [agents]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeployAgentRequest"
      responses:
        "201":
          description: Created member agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberWithAgent"
        "404":
          description: Source agent not found
        "409":
          description: Agent slug already exists in project
  /v1/projects/{project_id}/agents/{agent_slug}:
    delete:
      summary: Delete a project agent and its kernel AgentConfig
      operationId: deleteProjectAgent
      tags: [agents]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: agent_slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Agent not found
  /v1/agents/proposals/{session_id}/confirm:
    post:
      summary: Confirm a proposed agent (create in the library + deploy into the session's project)
      description: >-
        User confirmation of an agent the assistant proposed via the
        ``propose_agent`` tool. Creates a library agent and, when the session is
        bound to a project, deploys it into that project as a member. The frontend
        replays the agent spec from the originating ``tool_use`` event into the
        request body.
      operationId: confirmAgentProposal
      tags: [agents]
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposeAgentConfirmRequest"
      responses:
        "201":
          description: Agent created (and deployed when the session has a project)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProposeAgentConfirmResponse"
        "404":
          description: Session not found
        "409":
          description: Agent already exists
...
components:
  schemas:
...
    Agent:
      type: object
      required: [id, slug, name, description, instructions, runtime, model, skills, connector_types, source]
      properties:
        id: {type: string}
        slug: {type: string}
        name: {type: string}
        description: {type: string}
        instructions: {type: string}
        runtime: {type: string}
        model: {type: string}
        skills:
          type: array
          items: {type: string}
        connector_types:
          type: array
          items: {type: string}
        provider_id:
          type: string
          nullable: true
          description: >-
            Default model provider id for instances ((provider, model) pair).
            null for seeded/official agents (provider ids are install-local).
        effort:
          type: string
          nullable: true
          enum: [low, medium, high, xhigh, max]
          description: >-
            Default reasoning-effort budget prefilled into instances. null =
            no override (runtime SDK default). Project conversations read the
            budget from the bound agent.
        source: {type: string}
        readonly: {type: boolean, default: false, description: "If true, the agent cannot be edited."}
        deletable: {type: boolean, default: true, description: "If false, the agent cannot be deleted."}
        avatar:
          type: string
          nullable: true
          description: >-
            Preset icon key or uploaded asset URL for the agent's avatar.
            v1 supports preset keys only. null = no avatar set.
    CreateAgentRequest:
      type: object
      required: [name]
      properties:
        # Optional — backend derives a CJK-preserving, globally-unique slug
        # from ``name`` when omitted (VALUZ-AGENT-SLUG). UI sends name only.
        slug: {type: string, nullable: true}
        name: {type: string}
        description: {type: string, default: ""}
        instructions: {type: string, default: ""}
        runtime: {type: string, default: claude_agent}
        model: {type: string, default: claude-sonnet-4-6}
        skills:
          type: array
          items: {type: string}
        connector_types:
          type: array
          items: {type: string}
        provider_id: {type: string, nullable: true}
        effort: {type: string, nullable: true, enum: [low, medium, high, xhigh, max]}
        avatar: {type: string, nullable: true}
    UpdateAgentRequest:
      type: object
      properties:
        name: {type: string, nullable: true}
        description: {type: string, nullable: true}
        instructions: {type: string, nullable: true}
        runtime: {type: string, nullable: true}
        model: {type: string, nullable: true}
        skills:
          type: array
          nullable: true
          items: {type: string}
        connector_types:
          type: array
          nullable: true
          items: {type: string}
        provider_id: {type: string, nullable: true}
        effort: {type: string, nullable: true, enum: [low, medium, high, xhigh, max]}
        avatar: {type: string, nullable: true}
    ConnectorBindingInput:
      type: object
      required: [type]
      properties:
        type: {type: string}
        account_id: {type: string, nullable: true}
    CreateBlankAgentRequest:
      type: object
      required: [name]
      properties:
        # Optional — backend derives from ``name``, unique within the
        # project, when omitted (VALUZ-AGENT-SLUG).
        agent_slug: {type: string, nullable: true}
        name: {type: string}
        instructions: {type: string, default: ""}
        runtime: {type: string, default: claude_agent}
        model: {type: string, default: claude-sonnet-4-6}
        provider_id: {type: string, nullable: true}
        effort: {type: string, nullable: true, enum: [low, medium, high, xhigh, max]}
        skills:
          type: array
          nullable: true
          items: {type: string}
        connector_bindings:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ConnectorBindingInput"
    DeployAgentRequest:
      type: object
      required: [source_agent_slug]
      description: >-
        v2 派驻 (live reference). Config lives on the agent, so there is no
        per-deploy provider/model/connector override — just the source agent and
        an optional project-local handle.
      properties:
        source_agent_slug: {type: string}
        # Optional — backend derives from the source agent's name, unique
        # within the target project, when omitted (VALUZ-AGENT-SLUG).
        agent_slug: {type: string, nullable: true}
    ProjectMember:
      type: object
      required: [id, project_id, agent_slug]
      properties:
        id: {type: string}
        project_id: {type: string}
        agent_slug: {type: string}
        source_agent_slug: {type: string, nullable: true}
    AgentSummary:
      type: object
      required: [id, name, model, runtime_provider, instructions, skills, connectors]
      properties:
        id: {type: string}
        name: {type: string}
        model: {type: string}
        runtime_provider: {type: string}
        instructions: {type: string}
        skills:
          type: array
          items: {type: string}
        connectors:
          type: array
          items: {type: string}
        provider_id: {type: string, nullable: true}
        effort: {type: string, nullable: true, enum: [low, medium, high, xhigh, max]}
    MemberWithAgent:
      type: object
      required: [member, agent]
      properties:
        member:
          $ref: "#/components/schemas/ProjectMember"
        agent:
          oneOf:
            - $ref: "#/components/schemas/AgentSummary"
            - type: "null"
    ProposeAgentConfirmRequest:
      type: object
      description: >-
        Spec of an agent the user is confirming. Mirrors the ``propose_agent``
        tool input; the backend derives a unique slug from ``name``.
      required: [name, instructions]
      properties:
        name: {type: string}
        instructions: {type: string}
        description: {type: string, default: ""}
        runtime: {type: string, default: claude_agent}
        model: {type: string, default: claude-sonnet-4-6}
        effort: {type: string, nullable: true, enum: [low, medium, high, xhigh, max]}
        skills:
          type: array
          items: {type: string}
        connectors:
          type: array
          items: {type: string}
        avatar: {type: string, nullable: true}
    ProposeAgentConfirmResponse:
      type: object
      required: [agent, deployed]
      properties:
        agent:
          $ref: "#/components/schemas/AgentSummary"
        member:
          oneOf:
            - $ref: "#/components/schemas/ProjectMember"
            - type: "null"
        deployed: {type: boolean}
        project_id: {type: string, nullable: true}
...